ATR-414: refactored out constructors in DAC analyzer
